+++ internal/providers/upstashbox/core.go
@@ -20,6 +20,7 @@ type DoctorResult = core.DoctorResult
 type WarmupRequest = core.WarmupRequest
 type RunRequest = core.RunRequest
 type RunResult = core.RunResult
+type RunSessionHandle = core.RunSessionHandle
 type ListRequest = core.ListRequest
 type LeaseView = core.LeaseView
 type StatusRequest = core.StatusRequest
@@ -103,6 +104,10 @@ func printEnvForwardingSummary(w io.Writer, provider, behavior string, allow []s
 
 func shellQuote(s string) string {
 	return core.ShellQuote(s)
+}
+
+func upstashBoxCleanupCommand(leaseID string) string {
+	return "crabbox stop --provider " + providerName + " " + shellQuote(leaseID)
 }
 
 func shellScriptFromArgv(command []string) string {